+++ packages/stream_chat_flutter/lib/src/full_screen_media.dart
@@ -5,6 +5,7 @@ import 'package:cached_network_image/cached_network_image.dart';
 import 'package:chewie/chewie.dart';
 import 'package:flutter/material.dart';
 import 'package:photo_view/photo_view.dart';
+import 'package:shimmer/shimmer.dart';
 import 'package:stream_chat_flutter/src/extension.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 import 'package:video_player/video_player.dart';
@@ -64,36 +65,21 @@ class StreamFullScreenMedia extends StatefulWidget {
   _StreamFullScreenMediaState createState() => _StreamFullScreenMediaState();
 }
 
-class _StreamFullScreenMediaState extends State<StreamFullScreenMedia>
-    with SingleTickerProviderStateMixin {
-  late final AnimationController _animationController;
+class _StreamFullScreenMediaState extends State<StreamFullScreenMedia> {
   late final PageController _pageController;
 
-  late final _curvedAnimation = CurvedAnimation(
-    parent: _animationController,
-    curve: Curves.easeOut,
-    reverseCurve: Curves.easeIn,
-  );
+  late final _currentPage = ValueNotifier(widget.startIndex);
+  late final _isDisplayingDetail = ValueNotifier<bool>(true);
 
-  final _opacityTween = Tween<double>(begin: 1, end: 0);
-  late final _opacityAnimation = _opacityTween.animate(
-    CurvedAnimation(
-      parent: _animationController,
-      curve: const Interval(0, 0.6, curve: Curves.easeOut),
-    ),
-  );
-
-  late final ValueNotifier<int> _currentPage = ValueNotifier(widget.startIndex);
+  void switchDisplayingDetail() {
+    _isDisplayingDetail.value = !_isDisplayingDetail.value;
+  }
 
   final videoPackages = <String, VideoPackage>{};
 
   @override
   void initState() {
     super.initState();
-    _animationController = AnimationController(
-      vsync: this,
-      duration: const Duration(milliseconds: 300),
-    );
     _pageController = PageController(initialPage: widget.startIndex);
     for (var i = 0; i < widget.mediaAttachmentPackages.length; i++) {
       final attachment = widget.mediaAttachmentPackages[i].attachment;
@@ -127,26 +113,102 @@ class _StreamFullScreenMediaState extends State<StreamFullScreenMedia>
   @override
   Widget build(BuildContext context) => Scaffold(
         resizeToAvoidBottomInset: false,
-        body: Stack(
-          children: [
-            PageView.builder(
+        body: ValueListenableBuilder<int>(
+          valueListenable: _currentPage,
+          builder: (context, currentPage, child) {
+            final _currentAttachmentPackage =
+                widget.mediaAttachmentPackages[currentPage];
+            final _currentMessage = _currentAttachmentPackage.message;
+            final _currentAttachment = _currentAttachmentPackage.attachment;
+            return Stack(
+              children: [
+                Positioned.fill(child: child!),
+                ValueListenableBuilder<bool>(
+                  valueListenable: _isDisplayingDetail,
+                  builder: (context, isDisplayingDetail, child) {
+                    final mediaQuery = MediaQuery.of(context);
+                    final topPadding = mediaQuery.padding.top;
+                    return AnimatedPositionedDirectional(
+                      duration: kThemeAnimationDuration,
+                      curve: Curves.easeInOut,
+                      top: isDisplayingDetail
+                          ? 0
+                          : -(topPadding + kToolbarHeight),
+                      start: 0,
+                      end: 0,
+                      height: topPadding + kToolbarHeight,
+                      child: StreamGalleryHeader(
+                        userName: widget.userName,
+                        sentAt: context.translations.sentAtText(
+                          date: _currentAttachmentPackage.message.createdAt,
+                          time: _currentAttachmentPackage.message.createdAt,
+                        ),
+                        onBackPressed: Navigator.of(context).pop,
+                        message: _currentMessage,
+                        attachment: _currentAttachment,
+                        onShowMessage: () {
+                          widget.onShowMessage?.call(
+                            _currentMessage,
+                            StreamChannel.of(context).channel,
+                          );
+                        },
+                        attachmentActionsModalBuilder:
+                            widget.attachmentActionsModalBuilder,
+                      ),
+                    );
+                  },
+                ),
+                if (!_currentMessage.isEphemeral)
+                  ValueListenableBuilder<bool>(
+                    valueListenable: _isDisplayingDetail,
+                    builder: (context, isDisplayingDetail, child) {
+                      final mediaQuery = MediaQuery.of(context);
+                      final bottomPadding = mediaQuery.padding.bottom;
+                      return AnimatedPositionedDirectional(
+                        duration: kThemeAnimationDuration,
+                        curve: Curves.easeInOut,
+                        bottom: isDisplayingDetail
+                            ? 0
+                            : -(bottomPadding + kToolbarHeight),
+                        start: 0,
+                        end: 0,
+                        height: bottomPadding + kToolbarHeight,
+                        child: StreamGalleryFooter(
+                          currentPage: currentPage,
+                          totalPages: widget.mediaAttachmentPackages.length,
+                          mediaAttachmentPackages:
+                              widget.mediaAttachmentPackages,
+                          mediaSelectedCallBack: (val) {
+                            _currentPage.value = val;
+                            _pageController.animateToPage(
+                              val,
+                              duration: const Duration(milliseconds: 300),
+                              curve: Curves.easeInOut,
+                            );
+                            Navigator.pop(context);
+                          },
+                        ),
+                      );
+                    },
+                  ),
+              ],
+            );
+          },
+          child: InkWell(
+            onTap: switchDisplayingDetail,
+            child: PageView.builder(
               controller: _pageController,
+              itemCount: widget.mediaAttachmentPackages.length,
               onPageChanged: (val) {
                 _currentPage.value = val;
-
-                if (videoPackages.isEmpty) {
-                  return;
-                }
-
+                if (videoPackages.isEmpty) return;
                 final currentAttachment =
                     widget.mediaAttachmentPackages[val].attachment;
-
                 for (final e in videoPackages.values) {
                   if (e._attachment != currentAttachment) {
                     e._chewieController?.pause();
                   }
                 }
-
                 if (widget.autoplayVideos &&
                     currentAttachment.type == 'video') {
                   final controller = videoPackages[currentAttachment.id]!;
@@ -161,33 +223,44 @@ class _StreamFullScreenMediaState extends State<StreamFullScreenMedia>
                   final imageUrl = attachment.imageUrl ??
                       attachment.assetUrl ??
                       attachment.thumbUrl;
-                  return AnimatedBuilder(
-                    animation: _curvedAnimation,
-                    builder: (context, child) => PhotoView(
-                      loadingBuilder: (context, image) => const Offstage(),
-                      imageProvider: (imageUrl == null &&
-                              attachment.localUri != null &&
-                              attachment.file?.bytes != null)
-                          ? Image.memory(attachment.file!.bytes!).image
-                          : CachedNetworkImageProvider(imageUrl!),
-                      maxScale: PhotoViewComputedScale.covered,
-                      minScale: PhotoViewComputedScale.contained,
-                      heroAttributes: PhotoViewHeroAttributes(
-                        tag: widget.mediaAttachmentPackages,
+                  return ValueListenableBuilder<bool>(
+                    valueListenable: _isDisplayingDetail,
+                    builder: (context, isDisplayingDetail, _) =>
+                        AnimatedContainer(
+                      color: isDisplayingDetail
+                          ? StreamChannelHeaderTheme.of(context).color
+                          : Colors.black,
+                      duration: kThemeAnimationDuration,
+                      child: PhotoView(
+                        imageProvider: (imageUrl == null &&
+                                attachment.localUri != null &&
+                                attachment.file?.bytes != null)
+                            ? Image.memory(attachment.file!.bytes!).image
+                            : CachedNetworkImageProvider(imageUrl!),
+                        errorBuilder: (_, __, ___) => const AttachmentError(),
+                        loadingBuilder: (context, _) {
+                          final image = Image.asset(
+                            'images/placeholder.png',
+                            fit: BoxFit.cover,
+                            package: 'stream_chat_flutter',
+                          );
+                          final colorTheme =
+                              StreamChatTheme.of(context).colorTheme;
+                          return Shimmer.fromColors(
+                            baseColor: colorTheme.disabled,
+                            highlightColor: colorTheme.inputBg,
+                            child: image,
+                          );
+                        },
+                        maxScale: PhotoViewComputedScale.covered,
+                        minScale: PhotoViewComputedScale.contained,
+                        heroAttributes: PhotoViewHeroAttributes(
+                          tag: widget.mediaAttachmentPackages,
+                        ),
+                        backgroundDecoration: const BoxDecoration(
+                          color: Colors.transparent,
+                        ),
                       ),
-                      backgroundDecoration: BoxDecoration(
-                        color: ColorTween(
-                          begin: StreamChannelHeaderTheme.of(context).color,
-                          end: Colors.black,
-                        ).lerp(_curvedAnimation.value),
-                      ),
-                      onTapUp: (a, b, c) {
-                        if (_animationController.isCompleted) {
-                          _animationController.reverse();
-                        } else {
-                          _animationController.forward();
-                        }
-                      },
                     ),
                   );
                 } else if (attachment.type == 'video') {
@@ -198,17 +271,9 @@ class _StreamFullScreenMediaState extends State<StreamFullScreenMedia>
                     );
                   }
                   return InkWell(
-                    onTap: () {
-                      if (_animationController.isCompleted) {
-                        _animationController.reverse();
-                      } else {
-                        _animationController.forward();
-                      }
-                    },
+                    onTap: switchDisplayingDetail,
                     child: Padding(
-                      padding: const EdgeInsets.symmetric(
-                        vertical: 50,
-                      ),
+                      padding: const EdgeInsets.symmetric(vertical: 50),
                       child: Chewie(
                         controller: controller.chewieController!,
                       ),
@@ -217,76 +282,16 @@ class _StreamFullScreenMediaState extends State<StreamFullScreenMedia>
                 }
                 return const SizedBox();
               },
-              itemCount: widget.mediaAttachmentPackages.length,
             ),
-            FadeTransition(
-              opacity: _opacityAnimation,
-              child: ValueListenableBuilder<int>(
-                valueListenable: _currentPage,
-                builder: (context, value, child) {
-                  final _currentAttachmentPackage =
-                      widget.mediaAttachmentPackages[value];
-                  final _currentMessage = _currentAttachmentPackage.message;
-                  final _currentAttachment =
-                      _currentAttachmentPackage.attachment;
-                  return Column(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      StreamGalleryHeader(
-                        userName: widget.userName,
-                        sentAt: context.translations.sentAtText(
-                          date: widget
-                              .mediaAttachmentPackages[_currentPage.value]
-                              .message
-                              .createdAt,
-                          time: widget
-                              .mediaAttachmentPackages[_currentPage.value]
-                              .message
-                              .createdAt,
-                        ),
-                        onBackPressed: () {
-                          Navigator.of(context).pop();
-                        },
-                        message: _currentMessage,
-                        attachment: _currentAttachment,
-                        onShowMessage: () {
-                          widget.onShowMessage?.call(
-                            _currentMessage,
-                            StreamChannel.of(context).channel,
-                          );
-                        },
-                        attachmentActionsModalBuilder:
-                            widget.attachmentActionsModalBuilder,
-                      ),
-                      if (!_currentMessage.isEphemeral)
-                        StreamGalleryFooter(
-                          currentPage: value,
-                          totalPages: widget.mediaAttachmentPackages.length,
-                          mediaAttachmentPackages:
-                              widget.mediaAttachmentPackages,
-                          mediaSelectedCallBack: (val) {
-                            _currentPage.value = val;
-                            _pageController.animateToPage(
-                              val,
-                              duration: const Duration(milliseconds: 300),
-                              curve: Curves.easeInOut,
-                            );
-                            Navigator.pop(context);
-                          },
-                        ),
-                    ],
-                  );
-                },
-              ),
-            ),
-          ],
+          ),
         ),
       );
 
   @override
   void dispose() {
-    _animationController.dispose();
+    _currentPage.dispose();
     _pageController.dispose();
+    _isDisplayingDetail.dispose();
     for (final package in videoPackages.values) {
       package.dispose();
     }