+++ packages/stream_chat_flutter/lib/src/v4/message_input/stream_attachment_picker.dart
@@ -43,6 +43,10 @@ class StreamAttachmentPicker extends StatefulWidget {
       DefaultAttachmentTypes.video,
     ],
     this.customAttachmentTypes = const [],
+    this.attachmentThumbnailSize = const ThumbnailSize(400, 400),
+    this.attachmentThumbnailFormat = ThumbnailFormat.jpeg,
+    this.attachmentThumbnailQuality = 100,
+    this.attachmentThumbnailScale = 1,
   });
 
   /// True if the picker is open.
@@ -77,6 +81,25 @@ class StreamAttachmentPicker extends StatefulWidget {
   /// The list of custom attachment types that can be picked.
   final List<CustomAttachmentType> customAttachmentTypes;
 
+  /// Size of the attachment thumbnails.
+  ///
+  /// Defaults to (400, 400).
+  final ThumbnailSize attachmentThumbnailSize;
+
+  /// Format of the attachment thumbnails.
+  ///
+  /// Defaults to [ThumbnailFormat.jpeg].
+  final ThumbnailFormat attachmentThumbnailFormat;
+
+  /// The quality value for the attachment thumbnails.
+  ///
+  /// Valid from 1 to 100.
+  /// Defaults to 100.
+  final int attachmentThumbnailQuality;
+
+  /// The scale to apply on the [attachmentThumbnailSize].
+  final double attachmentThumbnailScale;
+
   /// Used to create a new copy of [StreamAttachmentPicker] with modified
   /// properties.
   StreamAttachmentPicker copyWith({
@@ -91,7 +114,11 @@ class StreamAttachmentPicker extends StatefulWidget {
     ValueChanged<bool>? onChangeInputState,
     ValueChanged<String>? onError,
     List<DefaultAttachmentTypes>? allowedAttachmentTypes,
-    List<CustomAttachmentType>? customAttachmentTypes = const [],
+    List<CustomAttachmentType>? customAttachmentTypes,
+    ThumbnailSize? attachmentThumbnailSize,
+    ThumbnailFormat? attachmentThumbnailFormat,
+    int? attachmentThumbnailQuality,
+    double? attachmentThumbnailScale,
   }) =>
       StreamAttachmentPicker(
         key: key ?? this.key,
@@ -109,6 +136,14 @@ class StreamAttachmentPicker extends StatefulWidget {
             allowedAttachmentTypes ?? this.allowedAttachmentTypes,
         customAttachmentTypes:
             customAttachmentTypes ?? this.customAttachmentTypes,
+        attachmentThumbnailSize:
+            attachmentThumbnailSize ?? this.attachmentThumbnailSize,
+        attachmentThumbnailFormat:
+            attachmentThumbnailFormat ?? this.attachmentThumbnailFormat,
+        attachmentThumbnailQuality:
+            attachmentThumbnailQuality ?? this.attachmentThumbnailQuality,
+        attachmentThumbnailScale:
+            attachmentThumbnailScale ?? this.attachmentThumbnailScale,
       );
 
   @override
@@ -360,6 +395,11 @@ class _StreamAttachmentPickerState extends State<StreamAttachmentPicker> {
                         },
                         allowedAttachmentTypes: widget.allowedAttachmentTypes,
                         customAttachmentTypes: widget.customAttachmentTypes,
+                        mediaThumbnailSize: widget.attachmentThumbnailSize,
+                        mediaThumbnailFormat: widget.attachmentThumbnailFormat,
+                        mediaThumbnailQuality:
+                            widget.attachmentThumbnailQuality,
+                        mediaThumbnailScale: widget.attachmentThumbnailScale,
                       ),
                     ),
                   ),
@@ -374,9 +414,11 @@ class _StreamAttachmentPickerState extends State<StreamAttachmentPicker> {
   void _addAssetAttachment(AssetEntity medium) async {
     final mediaFile = await medium.originFile;
     if (mediaFile == null) return;
+
     final tempDir = await getTemporaryDirectory();
 
     // TODO: Confirm that this max resolution is final
+    // Taken from https://getstream.io/chat/docs/flutter-dart/file_uploads/?language=dart#image-resizing
     const maxCDNImageResolution = 16800000;
     final imageResolution = medium.width * medium.height;
     File? cachedFile;
@@ -388,12 +430,13 @@ class _StreamAttachmentPickerState extends State<StreamAttachmentPicker> {
           updatedSize.width.floor(),
           updatedSize.height.floor(),
         ),
-        quality: 30, // TODO: investigate compressing all images
+        quality: 70, // TODO: investigate compressing all images
       );
-      cachedFile =
+      final file =
           await File('${tempDir.path}/${mediaFile.path.split('/').last}')
-              .create()
-            ..writeAsBytesSync(resizedImage!);
+              .create();
+      file.writeAsBytesSync(resizedImage!);
+      cachedFile = file;
     } else {
       cachedFile = await mediaFile
           .copy('${tempDir.path}/${mediaFile.path.split('/').last}');
@@ -457,6 +500,10 @@ class _PickerWidget extends StatefulWidget {
     required this.allowedAttachmentTypes,
     required this.customAttachmentTypes,
     required this.mediaListViewController,
+    this.mediaThumbnailSize = const ThumbnailSize(400, 400),
+    this.mediaThumbnailFormat = ThumbnailFormat.jpeg,
+    this.mediaThumbnailQuality = 100,
+    this.mediaThumbnailScale = 1,
   });
 
   final int filePickerIndex;
@@ -468,6 +515,10 @@ class _PickerWidget extends StatefulWidget {
   final List<DefaultAttachmentTypes> allowedAttachmentTypes;
   final List<CustomAttachmentType> customAttachmentTypes;
   final MediaListViewController mediaListViewController;
+  final ThumbnailSize mediaThumbnailSize;
+  final ThumbnailFormat mediaThumbnailFormat;
+  final int mediaThumbnailQuality;
+  final double mediaThumbnailScale;
 
   @override
   _PickerWidgetState createState() => _PickerWidgetState();
@@ -522,6 +573,10 @@ class _PickerWidgetState extends State<_PickerWidget> {
             selectedIds: widget.selectedMedias,
             onSelect: widget.onMediaSelected,
             controller: widget.mediaListViewController,
+            thumbnailSize: widget.mediaThumbnailSize,
+            thumbnailFormat: widget.mediaThumbnailFormat,
+            thumbnailQuality: widget.mediaThumbnailQuality,
+            thumbnailScale: widget.mediaThumbnailScale,
           );
         }